requirements.txt
streamlit==1.29.0
deep-translator==1.11.4
langdetect==1.0.9
python-dotenv==1.0.0
google-generativeai==0.3.2
requests==2.31.0
llama-index==0.9.15
llama-index-agent-openai==0.1.4
llama-index-llms-ollama==0.1.2